Restructured 'core/directives'
* Renamed all directive files from 'sd<name>' to '<name>Directive' to match 'apps' counterpart
* Created new module 'superdesk.core.directives' that contains the whole package.
