+++ scripts/core/directives/CheckDirective.js
