+++ scripts/core/directives/ConfirmDirective.js
