+++ scripts/core/directives/FiletypeIconDirective.js
