+++ scripts/core/directives/PasswordStrengthDirective.js
