+++ scripts/core/directives/PermissionsDirective.js
