+++ scripts/core/directives/SearchListDirective.js
