+++ scripts/core/directives/TypeaheadDirective.js
