+++ scripts/core/directives/WithParamsDirective.js
@@ -1,4 +1,4 @@
-export default angular.module('superdesk.core.links', [])
+export default angular.module('superdesk.core.directives.withParams', [])
     /**
      * sdWithParams manipulates href attribute to include current parameters.
      *
